requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
sse-starlette>=1.8.2

# OpenAI & Embeddings
openai>=1.0.0

# Vector Database
pinecone>=3.0.0

# Git Operations
GitPython>=3.1.0

# Environment Variables
python-dotenv>=1.0.0

# Tree-Sitter (Optional but recommended for better chunking)
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-typescript>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-html>=0.20.0
tree-sitter-css>=0.20.0
 